refactor: adjust folder/file structure
